+++ R/download-photon.R
@@ -2,14 +2,16 @@
 #' @description
 #' Download the photon executable from GitHub.
 #'
-#' @param path Path to a directory to store the executable.
+#' @param path Path to a directory to store the executable. Defaults to
+#' \code{tempdir()}.
 #' @param version Version tag of the photon release. If \code{NULL},
-#' downloads the latest known version. A list of all
+#' downloads the latest known version (`r PHOTON_VERSION`). A list of all
 #' releases can be found here: \url{https://github.com/komoot/photon/releases/}.
-#' Ignored if \code{jar} is given. If \code{NULL}, uses the latest known
-#' version (Currently: `r get_latest_photon()`).
+#' Ignored if \code{jar} is given.
 #' @param opensearch If \code{TRUE}, downloads the OpenSearch version of
 #' photon if available. OpenSearch versions are available for photon >= 0.6.0.
+#' Since photon >= 0.7.0, OpenSearch versions are recommended. Defaults to
+#' \code{TRUE}.
 #' @inheritParams download_searchindex
 #'
 #' @returns If \code{only_url = FALSE}, returns a character string giving the
@@ -18,10 +20,10 @@
 #' @export
 #'
 #' @examples
-#' \donttest{download_photon(tempdir(), version = "0.4.1")}
-download_photon <- function(path = ".",
+#' \donttest{download_photon(tempdir(), version = "0.4.1", opensearch = FALSE)}
+download_photon <- function(path = tempdir(),
                             version = NULL,
-                            opensearch = FALSE,
+                            opensearch = TRUE,
                             only_url = FALSE,
                             quiet = FALSE) {
   assert_dir(path)
@@ -39,10 +41,11 @@ download_photon <- function(path = ".",
   }
 
   if (!quiet) {
+    type <- ifelse(opensearch, "OpenSearch", "ElasticSearch")
     cli::cli_progress_step(
-      msg = "Fetching photon {.field {version}}.",
-      msg_done = "Successfully downloaded photon {.field {version}}.",
-      msg_failed = "Failed to download photon."
+      msg = "Fetching {type} photon {.field {version}}.",
+      msg_done = "Successfully downloaded {type} photon {.field {version}}.",
+      msg_failed = "Failed to download {type} photon."
     )
   }