+++ R/new_photon.R
@@ -21,7 +21,9 @@
 #' version.
 #' @param country Character string that can be identified by
 #' \code{\link[countrycode]{countryname}} as a country. An extract for this
-#' country will be downloaded. If \code{NULL}, downloads a global search index.
+#' country will be downloaded. If \code{"planet"}, downloads a global search
+#' index. If \code{NULL}, downloads no index and leaves download or import to
+#' the user.
 #' @param date Character string or date-time object used to specify the creation
 #' date of the search index. If \code{"latest"}, will download the file tagged
 #' with "latest". If a character string, the value should be parseable by
@@ -29,11 +31,13 @@
 #' compared to all available dates and the closest date will be selected.
 #' Otherwise, a file will be selected that exactly matches the input to
 #' \code{date}.
-#' @param opensearch If \code{TRUE}, looks for an OpenSearch version of
-#' photon in the specified path. Opensearch-based photon supports structured
-#' geocoding queries but is currently only experimental. Defaults to
-#' \code{FALSE}. See \code{vignette("nominatim-import", package = "photon")}
-#' for details.
+#' @param opensearch If \code{TRUE}, attempts to download the OpenSearch
+#' version of photon. OpenSearch-based photon supports structrued geocoding.
+#' Readily available OpenSearch photon executables are only offered since
+#' photon version 0.6.0. For earlier versions, you need to build it from
+#' source using gradle. In this case, if \code{TRUE}, will look for an
+#' OpenSearch version of photon in the specified path. Since photon version
+#' 0.7.0, OpenSearch is the recommended option. Defaults to \code{TRUE}.
 #' @param exact If \code{TRUE}, exactly matches the \code{date}. Otherwise,
 #' selects the date with lowest difference to the \code{date} parameter.
 #' @param section Subdirectory of the download server from which to select a
@@ -41,6 +45,10 @@
 #' version of photon. If \code{"archived"}, selects a dump made for an older
 #' version of photon. If \code{NULL} (or any arbitrary string), selects a
 #' dump made for the current release. Defaults to \code{NULL}.
+#' @param mount If \code{TRUE}, mounts the object to the session so that
+#' functions like \code{\link{geocode}} automatically detect the new
+#' instance. If \code{FALSE}, initializies the instance but doesn't mount
+#' it to the session. Defaults to \code{TRUE}.
 #' @param overwrite If \code{TRUE}, overwrites existing jar files and
 #' search indices when initializing a new instance. Defaults to
 #' \code{FALSE}.
@@ -55,12 +63,15 @@
 #' photon <- new_photon()
 #'
 #' # connect to arbitrary server
-#' photon <- new_photon(url = "photonserver.org")
+#' photon <- new_photon(url = "https://photonserver.org")
 #'
 #' if (has_java("11")) {
-#' # set up a local instance in the current working directory
-#' photon <- new_photon("photon", country = "Monaco")
+#' # set up a local instance in a temporary directory
+#' dir <- file.path(tempdir(), "photon")
+#' photon <- new_photon(dir, country = "Monaco")
 #' }
+#'
+#' photon$purge(ask = FALSE)
 new_photon <- function(path = NULL,
                        url = NULL,
                        photon_version = NULL,
@@ -68,13 +79,14 @@ new_photon <- function(path = NULL,
                        date = "latest",
                        exact = FALSE,
                        section = NULL,
-                       opensearch = FALSE,
+                       opensearch = TRUE,
+                       mount = TRUE,
                        overwrite = FALSE,
                        quiet = FALSE) {
   if (is.null(path) && is.null(url)) {
-    photon_remote$new(url = "https://photon.komoot.io/")
+    photon_remote$new(url = "https://photon.komoot.io/", mount = mount)
   } else if (is.null(path)) {
-    photon_remote$new(url = url)
+    photon_remote$new(url = url, mount = mount)
   } else if (is.null(url)) {
     photon_local$new(
       path = path,
@@ -84,6 +96,7 @@ new_photon <- function(path = NULL,
       exact = exact,
       section = section,
       opensearch = opensearch,
+      mount = mount,
       overwrite = overwrite,
       quiet = quiet
     )
@@ -96,6 +109,39 @@ photon_env <- new.env(parent = emptyenv())
 
 photon_cache <- function() {
   get("photon_env", envir = asNamespace("photon"))
+}
+
+
+#' Local photon instances
+#' @description
+#' Evaluate R code with a photon instance without changing the active photon
+#' mount.
+#'
+#' @param photon An object of class \code{\link[=new_photon]{photon}} that is
+#' temporarily mounted to the session.
+#' @param code Code to execute in the temporary environment.
+#'
+#' @returns The results of the evaluation of the \code{code} argument.
+#'
+#' @export
+#'
+#' @examples
+#' # Get a public instance
+#' pub_photon <- new_photon()
+#'
+#' # Mount a custom instance
+#' new_photon(url = "https://localhost:8001/")
+#'
+#' # Geocode with the public instance only once
+#' with_photon(pub_photon, geocode("Rutland"))
+#'
+#' # The custom instance is still mounted
+#' get_instance()
+with_photon <- function(photon, code) {
+  old <- get0("instance", envir = photon_cache())
+  on.exit(assign("instance", old, envir = photon_cache()))
+  photon$mount()
+  force(code)
 }