+++ R/photon_local.R
@@ -2,25 +2,27 @@
 #' @description
 #' This R6 class is used to initialize and manage local photon instances.
 #' It can download and setup the Java, the photon executable, and the necessary
-#' ElasticSearch search index. It can start, stop, and query the status of the
-#' photon instance. It is also the basis for geocoding requests at it is used
+#' OpenSearch index. It can start, stop, and query the status of the
+#' photon instance. It is also the basis for geocoding requests as it is used
 #' to retrieve the URL for geocoding.
 #'
-#' @section ElasticSearch / OpenSearch:
-#' The standard version of photon uses ElasticSearch indices to geocode.
-#' These search indices can be self-provided by importing an existing
+#' @section Search indices:
+#' Search indices can be self-provided by importing an existing
 #' Nominatim database or they can be downloaded from the
 #' \href{https://nominatim.org/2020/10/21/photon-country-extracts.html}{Photon download server}.
-#' Use \code{nominatim = TRUE} to indicate that no ElasticSearch indices
-#' should be downloaded. See
+#' If you want to download pre-built search indices, simply provide a
+#' \code{country} string during initialization or use the
+#' \code{$download_data} method. Pre-built search indices do not come with
+#' support for structured geocoding.
+#'
+#' If you want to build from Nominatim, do not
+#' provide a country string and use the \code{$import} method. See
 #' \code{vignette("nominatim-import", package = "photon")} for details on how
 #' to import from Nominatim.
 #'
 #' To enable structured geocoding, the photon geocoder needs to be built to
-#' support OpenSearch. Since photon 0.6.0, OpenSearch jar files are included
-#' in the photon releases. OpenSearch indices can also be downloaded, but
-#' do not support structured geocoding as of yet. To enable structured
-#' geocoding, indices have to be imported from an existing Nominatim database.
+#' support OpenSearch. Since photon 0.7.0, OpenSearch jar files are the
+#' standard and ElasticSearch is deprecated.
 #'
 #' @export
 #' @import R6
@@ -33,15 +35,17 @@
 #' photon <- new_photon(path = dir, country = "Monaco")
 #'
 #' # start a new instance with an older photon version
-#' photon <- new_photon(path = dir, photon_version = "0.4.1")}
+#' photon <- new_photon(path = dir, photon_version = "0.4.1", opensearch = FALSE)
+#' }
 #'
 #' \dontrun{
 #' # import a nominatim database using OpenSearch photon
 #' # this example requires the OpenSearch version of photon and a running
 #' # Nominatim server.
 #' photon <- new_photon(path = dir, opensearch = TRUE)
-#' photon$start(photon_options = cmd_options(port = 29146, password = "pgpass"))}
-#' }
+#' photon$import(photon_options = cmd_options(port = 29146, password = "pgpass"))}
+#'
+#' photon$purge(ask = FALSE)}
 photon_local <- R6::R6Class(
   inherit = photon,
   classname = "photon_local",
@@ -82,12 +86,18 @@ photon_local <- R6::R6Class(
     #' version of photon. If \code{"archived"}, selects a dump made for an older
     #' version of photon. If \code{NULL} (or any arbitrary string), selects a
     #' dump made for the current release. Defaults to \code{NULL}.
-    #' @param opensearch If \code{TRUE}, looks for an OpenSearch version of
-    #' photon in the specified path. Opensearch-based photon supports structured
-    #' geocoding queries but has to be built manually using gradle. Hence,
-    #' it cannot be downloaded directly. If no OpenSearch executable is found
-    #' in the search path, then this parameter is set to \code{FALSE}. Defaults
-    #' to \code{FALSE}.
+    #' @param opensearch Superseded for photon versions >= 0.7.0. If \code{TRUE},
+    #' attempts to download the OpenSearch version of photon. OpenSearch-based
+    #' photon supports structrued geocoding. Readily available OpenSearch
+    #' photon executables are only offered since photon version 0.6.0. For
+    #' earlier versions, you need to build from source using gradle. In this
+    #' case, if \code{TRUE}, will look for an OpenSearch version of photon in
+    #' the specified path. Since photon version 0.7.0, OpenSearch is the
+    #' recommended option. Defaults to \code{TRUE}.
+    #' @param mount If \code{TRUE}, mounts the object to the session so that
+    #' functions like \code{\link{geocode}} automatically detect the new
+    #' instance. If \code{FALSE}, initializies the instance but doesn't mount
+    #' it to the session. Defaults to \code{TRUE}.
     #' @param overwrite If \code{TRUE}, overwrites existing jar files and
     #' search indices when initializing a new instance. Defaults to
     #' \code{FALSE}.
@@ -97,14 +107,16 @@ photon_local <- R6::R6Class(
                           country = NULL,
                           date = "latest",
                           exact = FALSE,
-                          section = NULL,
-                          opensearch = FALSE,
+                          section = "experimental",
+                          opensearch = TRUE,
+                          mount = TRUE,
                           overwrite = FALSE,
                           quiet = FALSE) {
       assert_flag(quiet)
       assert_flag(opensearch)
       check_jdk_version("11", quiet = quiet)
       photon_version <- photon_version %||% get_latest_photon()
+      check_opensearch(opensearch, photon_version)
 
       path <- normalizePath(path, "/", mustWork = FALSE)
       if (!dir.exists(path)) {
@@ -131,7 +143,7 @@ photon_local <- R6::R6Class(
       meta <- private$get_metadata(quiet = quiet)
       private$country <- meta$country
       private$date <- meta$date
-      self$mount()
+      if (mount) self$mount()
       invisible(self)
     },
 
@@ -145,12 +157,21 @@ photon_local <- R6::R6Class(
 
     #' @description
     #' Retrieve metadata about the java and photon version used as well
-    #' as the country and creation date of the Eleasticsearch search index.
+    #' as the country and creation date of the search index.
     #' @return A list containing the java version, the photon version, and
     #' if applicable, the spatial and temporal coverage of the search index.
     info = function() {
       info <- list(java = get_java_version(quiet = TRUE))
       c(info, private$get_metadata(quiet = TRUE))
+    },
+
+    #' @description
+    #' Print the default arguments to the R console. This can be helpful to
+    #' get a list of additional photon arguments for \code{$start()} or
+    #' \code{$import()}.
+    #' @return Nothing, but prints to the console.
+    help = function() {
+      cat(run_photon(self, private, mode = "help", photon_opts = "-h")$stdout)
     },
 
     #' @description
@@ -160,7 +181,7 @@ photon_local <- R6::R6Class(
     #' instance.
     #' @return \code{NULL}, invisibly.
     purge = function(ask = TRUE) {
-      if (interactive() || !ask) {
+      if (interactive() && ask) {
         cli::cli_inform(c("i" = paste( # nocov start
           "Purging an instance kills the photon process",
           "and removes the photon directory."
@@ -213,11 +234,11 @@ photon_local <- R6::R6Class(
     #' \code{FALSE}.
     #' @param timeout Time in seconds before the java process aborts. Defaults
     #' to 60 seconds.
-    #' @param java_opts List of further flags passed on to the \code{java}
-    #' command.
-    #' @param photon_opts List of further flags passed on to the photon
-    #' jar in the java command. See \code{\link{cmd_options}} for a helper
-    #' function to import external Nominatim databases.
+    #' @param java_opts Character vector of further flags passed on to the
+    #' \code{java} command.
+    #' @param photon_opts Character vector of further flags passed on to the
+    #' photon jar in the java command. See \code{\link{cmd_options}} for a
+    #' helper function.
     import = function(host = "127.0.0.1",
                       port = 5432,
                       database = "nominatim",
@@ -249,12 +270,12 @@ photon_local <- R6::R6Class(
       assert_flag(json)
 
       if (structured && !private$opensearch) {
-        cli::cli_warn(paste(
+        cli::cli_warn(paste( # nocov start
           "Structured queries are only supported for OpenSearch photon.",
           "Setting {.code structured = FALSE}."
         ), class = "structured_elasticsearch_error")
         structured <- FALSE
-      }
+      } # nocov end
 
       popts <- cmd_options(
         nominatim_import = TRUE,
@@ -294,18 +315,17 @@ photon_local <- R6::R6Class(
     #' Defaults to \code{FALSE}.
     #' @param timeout Time in seconds before the java process aborts. Defaults
     #' to 60 seconds.
-    #' @param java_opts List of further flags passed on to the \code{java}
-    #' command.
-    #' @param photon_opts List of further flags passed on to the photon
-    #' jar in the java command. See \code{\link{cmd_options}} for a helper
-    #' function to import external Nominatim databases.
+    #' @param java_opts Character vector of further flags passed on to the
+    #' \code{java} command.
+    #' @param photon_opts Character vector of further flags passed on to the
+    #' photon jar in the java command. See \code{\link{cmd_options}} for a
+    #' helper function.
     #'
     #' @details
     #' While there is a certain way to determine if a photon instance is
     #' ready, there is no clear way as of yet to determine if a photon setup
-    #' has failed. Due to this, a failing setup is mostly indicated by the
-    #' setup hanging after emitting a warning. In this case, the setup has to
-    #' be interrupted manually.
+    #' has failed. Due to this, a failing setup may sometimes hang instead of
+    #' emitting an error. In this case, please open a bug report.
     start = function(host = "0.0.0.0",
                      port = "2322",
                      ssl = FALSE,
@@ -359,7 +379,8 @@ photon_local <- R6::R6Class(
     #' Downloads a search index using \code{\link{download_searchindex}}.
     #' @param country Character string that can be identified by
     #' \code{\link[countrycode]{countryname}} as a country. An extract for this
-    #' country will be downloaded. If \code{NULL}, downloads a global search index.
+    #' country will be downloaded. If \code{"planet"}, downloads a global search
+    #' index.
     #' @param date Character string or date-time object used to specify the creation
     #' date of the search index. If \code{"latest"}, will download the file tagged
     #' with "latest". If a character string, the value should be parseable by
@@ -374,11 +395,11 @@ photon_local <- R6::R6Class(
     #' version of photon. If \code{"archived"}, selects a dump made for an older
     #' version of photon. If \code{NULL} (or any arbitrary string), selects a
     #' dump made for the current release. Defaults to \code{NULL}.
-    download_data = function(country = NULL,
+    download_data = function(country,
                              date = "latest",
                              exact = FALSE,
-                             section = NULL) {
-      download_searchindex( # nocov start
+                             section = "experimental") {
+      archive_path <- download_searchindex(
         path = self$path,
         quiet = private$quiet,
         country = country,
@@ -386,7 +407,9 @@ photon_local <- R6::R6Class(
         exact = exact,
         section = section
       )
-      invisible(self) # nocov end
+      untar_index(archive_path, self$path)
+      store_index_metadata(self$path, archive_path)
+      invisible(self)
     },
 
     #' @description
@@ -552,24 +575,24 @@ setup_photon_directory <- function(path,
       archive_path <- files[has_archive] # nocov
     }
 
-    untar_es_index(archive_path, path)
-    store_searchindex_metadata(path, archive_path)
+    untar_index(archive_path, path)
+    store_index_metadata(path, archive_path)
   } else if (!quiet) {
     inform_index_exists()
   }
 }
 
 
-untar_es_index <- function(archive_path, path) {
+untar_index <- function(archive_path, path) {
   untared <- utils::untar(archive_path, files = "photon_data", exdir = path)
 
   if (!identical(untared, 0L)) {
-    ph_stop("Failed to untar the ElasticSearch index.") # nocov
+    ph_stop("Failed to untar search index.") # nocov
   }
 }
 
 
-store_searchindex_metadata <- function(path, archive_path) {
+store_index_metadata <- function(path, archive_path) {
   meta <- utils::strcapture(
     pattern = "photon-db-?([a-z]{2})?-([0-9]+|latest)\\.tar\\.bz2",
     x = basename(archive_path),
@@ -626,6 +649,17 @@ construct_jar <- function(version = NULL, opensearch = FALSE) {
   version <- version %||% get_latest_photon()
   opensearch <- ifelse(opensearch, "-opensearch", "")
   sprintf("photon%s-%s.jar", opensearch, version)
+}
+
+
+check_opensearch <- function(opensearch, version) {
+  version <- numeric_version(version)
+  if (version >= "0.7.0" && !opensearch) {
+    cli::cli_warn(c( # nocov start
+      "!" = "ElasticSearch versions are superseded for photon 0.7.0 or higher.",
+      "i" = "You can set opensearch = TRUE, to use OpenSearch photon instead."
+    )) # nocov end
+  }
 }