+++ tests/testthat/test-geocode.R
@@ -13,8 +13,9 @@ test_that("basic requests work", {
   res3 <- geocode("Berlin", bbox = c(xmin = 0, xmax = 13, ymin = 52, ymax = 53))
   expect_failure(expect_equal(res1, res3))
 
-  res4 <- geocode("Berlin", locbias = c(10, 52), zoom = 12, locbias_scale = 0.1)
-  expect_failure(expect_equal(res1, res4))
+  res4a <- geocode("Berlin", locbias = c(-45, 30), zoom = 12, locbias_scale = 0.1, limit = 5)
+  res4b <- geocode("Berlin", limit = 5)
+  expect_failure(expect_equal(res4a, res4b))
 
   res5 <- geocode("notarealplace")
   expect_equal(nrow(res5), 1)
@@ -47,7 +48,7 @@ test_that("basic reversing works", {
 })
 
 test_that("reversing with sf works", {
-  sf <- sf::st_sfc(sf::st_point(c(8, 52)))
+  sf <- sf::st_sfc(sf::st_point(c(8, 52)), crs = 4326)
   res <- reverse(sf)
   expect_s3_class(res, "sf")
   expect_equal(nrow(res), 1)