+++ tests/testthat/test-setup.R
@@ -4,8 +4,8 @@ test_that("remote photons work", {
   photon <- new_photon()
   expect_true(is_komoot(photon$get_url()))
   expect_error(structured(), regexp = "disabled")
-  photon <- new_photon(url = "test.org")
-  expect_equal(photon$get_url(), "test.org")
+  photon <- new_photon(url = "https://test.org")
+  expect_equal(photon$get_url(), "https://test.org")
 })
 
 test_that("java checks work", {
@@ -46,23 +46,28 @@ test_that("logs can be parsed", {
   expect_true(sum(vapply(logs, FUN.VALUE = logical(1), \(x) all(is.na(x)))) == 1)
 })
 
-skip_if_offline("graphhopper.com")
-skip_on_cran()
-
 test_that("search indices are matched", {
   de_latest <- download_searchindex(only_url = TRUE, country = "Germany")
   expect_equal(basename(de_latest), "photon-db-de-latest.tar.bz2")
   global_latest <- download_searchindex(only_url = TRUE, country = "planet")
   expect_equal(basename(global_latest), "photon-db-latest.tar.bz2")
-  global_time <- download_searchindex(only_url = TRUE, date = Sys.Date())
+  global_time <- download_searchindex(only_url = TRUE, date = Sys.Date(), country = "Monaco")
   expect_match(basename(global_time), "photon-db-mc-[0-9]+\\.tar\\.bz2")
   expect_error(
-    download_searchindex(only_url = TRUE, date = Sys.Date(), exact = TRUE),
+    download_searchindex(only_url = TRUE, date = Sys.Date(), exact = TRUE, country = "Monaco"),
     class = "no_index_match"
   )
   expect_error(
     download_searchindex(only_url = TRUE, country = "not a country"),
     class = "country_invalid"
+  )
+})
+
+test_that("search index download signals a useful error", {
+  skip_if_offline("graphhopper.com")
+  expect_error(
+    download_searchindex(country = "Vatican"),
+    regexp = "Vatican City is not available"
   )
 })
 
@@ -73,73 +78,106 @@ test_that("opensearch is denied when unsupported", {
   )
 })
 
+skip_on_cran()
 skip_if_offline("github.com")
+skip_if_offline("graphhopper.com")
 skip_if_not(has_java("11"))
 
-test_that("local setup works", {
+describe("photon_local", {
   options(photon_setup_warn = FALSE)
+  on.exit(options(photon_setup_warn = NULL), add = TRUE)
   dir <- file.path(tempdir(), "photon")
-  photon <- new_photon(path = dir, country = "samoa")
-  on.exit({
-    options(photon_setup_warn = NULL)
-    photon$purge(ask = FALSE)
+  photon <- new_photon(path = dir, country = "monaco")
+  on.exit(photon$purge(ask = FALSE), add = TRUE)
+
+  it("can print", {
+    expect_no_error(print(photon))
   })
 
-  # test pre-setup
-  expect_no_error(print(photon))
-  photon <- new_photon(path = dir, country = "samoa")
-  expect_no_message(new_photon(path = dir, quiet = TRUE))
-  expect_error(photon$get_url(), class = "no_url_yet")
-  expect_error(
-    get_photon_executable(photon$path, get_latest_photon(), TRUE),
-    regexp = "could not be found"
-  )
-  expect_false(photon$is_ready())
-  expect_false(photon$is_running())
+  photon <- new_photon(path = dir, country = "monaco")
 
-  # test setup
-  photon$start(host = "127.0.0.1")
-  expect_true(photon$is_running())
-  expect_gt(nrow(geocode("Apai")), 0)
-  expect_error(photon$start(host = "127.0.0.1"), class = "photon_already_running")
+  it("can suppress verbosity", {
+    expect_no_message(new_photon(path = dir, quiet = TRUE))
+  })
 
-  with_mocked_bindings(
-    unlink = \(...) 1L,
-    .package = "base",
-    expect_warning(photon$remove_data(), class = "photon_data_not_removed")
-  )
+  it("communicates if not yet set up", {
+    expect_error(photon$get_url(), class = "no_url_yet")
+    expect_error(
+      get_photon_executable(photon$path, "0.1.0", FALSE),
+      regexp = "could not be found"
+    )
+    expect_false(photon$is_ready())
+    expect_false(photon$is_running())
+  })
 
-  photon$stop()
-  expect_false(photon$is_running())
+  it("can start", {
+    photon$start(host = "127.0.0.1")
+    expect_true(photon$is_running())
+    expect_false(anyNA(geocode("Monte Carlo")$osm_id))
+    expect_error(photon$start(host = "127.0.0.1"), class = "photon_already_running")
+  })
 
-  # test error handling
+  it("warns if data cannot be removed", {
+    with_mocked_bindings(
+      unlink = \(...) 1L,
+      .package = "base",
+      expect_warning(photon$remove_data(), class = "photon_data_not_removed")
+    )
+  })
+
+  it("can stop", {
+    photon$stop()
+    expect_false(photon$is_running())
+  })
+
   fail_dir <- file.path(tempdir(), "photon_fail")
   photon <- new_photon(path = fail_dir)
-  expect_error(photon$import(), class = "import_error")
-  logs <- photon$get_logs()
-  expect_s3_class(logs, "data.frame")
-  expect_contains(logs$type, "ERROR")
-  expect_contains(names(logs), "rid")
 
-  expect_error(photon$start(), class = "start_error")
-  logs <- photon$get_logs()
-  expect_equal(unique(logs$rid), c(1, 2))
+  it("intercepts start errors correctly", {
+    expect_error(photon$start(), class = "start_error")
+    logs <- photon$get_logs()
+    expect_s3_class(logs, "data.frame")
+    expect_contains(logs$type, "ERROR")
+    expect_contains(names(logs), "rid")
+  })
 
-  options(photon_setup_warn = TRUE)
-  expect_warning(expect_error(photon$start(photon_opts = "-notanoption"), class = "start_error"))
-  logs <- photon$get_logs()
-  expect_contains(logs$type, c("WARN", "ERROR"))
-  expect_match(logs$msg, "usage error", all = FALSE)
-  expect_equal(unique(logs$rid), c(1, 2, 3))
+  it("intercepts import errors correctly", {
+    expect_error(photon$import(), class = "import_error")
+    logs <- photon$get_logs()
+    expect_equal(unique(logs$rid), c(1, 2))
+  })
 
-  options(photon_setup_warn = FALSE)
-  expect_warning(
-    expect_error(photon$import(structured = TRUE)),
-    class = "structured_elasticsearch_error"
-  )
+  it("intercepts usage errors correctly", {
+    options(photon_setup_warn = TRUE)
+    on.exit(options(photon_setup_warn = NULL), add = TRUE)
+    expect_warning(expect_error(photon$start(photon_opts = "-notanoption"), class = "start_error"))
+    logs <- photon$get_logs()
+    expect_contains(logs$type, c("WARN", "ERROR"))
+    expect_match(logs$msg, "usage error", all = FALSE)
+    expect_equal(unique(logs$rid), c(1, 2, 3))
+  })
 
-  photon$stop()
-  expect_no_warning(photon$remove_data())
-  expect_false(dir.exists(file.path(photon$path, "photon_data")))
+  it("can remove index data", {
+    photon$stop()
+    expect_no_warning(photon$remove_data())
+    expect_false(dir.exists(file.path(photon$path, "photon_data")))
+  })
+
+  it("can run help", {
+    expect_output(photon$help(), regexp = "Usage: <main class>")
+  })
+
+  it("can download data manually", {
+    photon$remove_data()
+    photon$download_data("monaco")
+    photon$start(host = "127.0.0.1")
+    photon$stop()
+  })
 })
 
+
+test_that("with_photon works", {
+  photon1 <- new_photon(url = "https://test.org")
+  photon2 <- new_photon()
+  expect_error(with_photon(photon1, geocode("Berlin")), regexp = "Could not resolve")
+})