+++ NEWS.md
@@ -1,3 +1,28 @@
+# photon 0.7.4
+
+* Increment photon version to 0.7.4
+* Match package version to photon version
+* Update examples to work in photon 0.7.4
+* Set photon 0.7.4 as default and add a supersede warning if version < 0.7.0
+* Add photon type (OpenSearch/ElasticSearch) to error message
+* Adjust geocoding error detection to OpenSearch
+* Update vignettes to OpenSearch
+* Switch from Samoa to Monaco as example country (more reliable)
+* Add `mount` argument to `new_photon()`. If `FALSE`, instance is created but not mounted.
+* Add function `with_photon()` to execute code using a local photon instance
+* Add extra info to HTTP404 when search index download does not yield a result
+* Add `$help()` method to show raw argument information from jar file
+* Add CRS checks and transformations when an sf geometry is provided to `reverse()`
+* Improve URL checker by relying on `httr2::url_parse()`
+* Fix `$download_data()` method not untaring archive and storing metadata
+* Fix error detection during setup when encountering an exception without a timestamp
+* Fix error detection not recognizing OpenSearch import errors
+* Fix path arguments defaulting to `"."`
+* Fix an example in `new_photon()`
+* Fix progress bar in `reverse()`
+* Purge photon instances after examples
+
+
 # photon 0.3.5
 
 * Set `limit = 1` as default (#2)